feat(chips-area): add separate theme + grid sample and chips-area changes
