+++ projects/igniteui-angular/chips/src/chips/chips-area/chips-area.component.html
