+++ projects/igniteui-angular/chips/src/chips/chips-area/chips-area.component.scss
@@ -1,0 +1,2 @@
+@use 'themes/base';
+@use 'themes/shared';
