+++ projects/igniteui-angular/chips/src/chips/chips-area/themes/shared/_index.scss
@@ -1,0 +1,1 @@
+@forward 'indigo';
