+++ projects/igniteui-angular/chips/src/chips/chips-area/themes/shared/_indigo.scss
@@ -1,0 +1,7 @@
+@use 'igniteui-theming/sass/typography' as *;
+@use 'styles/themes/standalone' as *;
+
+@include themed-block(igx-chip-area, indigo) {
+    padding: rem(4px);
+    gap: rem(8px)
+}
