+++ projects/igniteui-angular/chips/src/chips/public_api.ts
@@ -1,9 +1,9 @@
 import { IgxPrefixDirective, IgxSuffixDirective } from 'igniteui-angular/input-group';
 import { IgxChipComponent } from './chip.component';
-import { IgxChipsAreaComponent } from './chips-area.component';
+import { IgxChipsAreaComponent } from './chips-area/chips-area.component';
 
 export * from './chip.component';
-export * from './chips-area.component';
+export * from './chips-area/chips-area.component';
 
 /* NOTE: Chips directives collection for ease-of-use import in standalone components scenario */
 export const IGX_CHIPS_DIRECTIVES = [
