+++ src/app/grid-auto-size/grid-auto-size.sample.scss
@@ -5,3 +5,8 @@
 	height: 100%;
 	align-items: start;
 }
+
+button {
+	margin-inline-end: 0.5rem;
+	margin-block-end: 0.5rem;
+}
